Projet TS structuré et stable
