+++ package.json
@@ -2,14 +2,13 @@
   "name": "@natchi/discordjs-handler",
   "version": "1.2.1",
   "description": "Module qui permet d'exécuter des événements et des commandes pour discord.js en handler",
-  "main": "./lib/index.js",
   "repository": "https://github.com/Natchi59/discordjs-handler.git",
   "author": "Natchi <nathan.caron@natchi.fr>",
   "license": "MIT",
-  "scripts": {
-    "devts": "tsnd --respawn --transpile-only ./test/index.ts",
-    "devjs": "nodemon ./test/index.js"
-  },
+  "main": "./lib/index.js",
+  "files": [
+    "lib"
+  ],
   "keywords": [
     "discordjs",
     "discord.js",
@@ -20,10 +19,15 @@
   "dependencies": {
     "discord.js": "^12.5.3"
   },
+  "engines": {
+    "node": ">=14.0.0",
+    "npm": ">=7.0.0"
+  },
+  "scripts": {
+    "build": "tsc"
+  },
   "devDependencies": {
-    "@types/node": "^16.3.3",
-    "nodemon": "^2.0.12",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5"
+    "@types/node": "^16.4.0",
+    "@types/ws": "^7.4.7"
   }
 }
