+++ src/index.ts
@@ -1,0 +1,2 @@
+export { ClientHandler } from "./client";
+export { handlerEvents } from "./handlers/handlerEvents";
