+++ src/interfaces/Event.ts
@@ -1,0 +1,6 @@
+import { Client, ClientEvents } from "discord.js";
+
+export interface Event {
+  name: keyof ClientEvents;
+  run (client: Client, ...args: Array<any>): Promise<any> | any;
+}
