+++ src/interfaces/Options.ts
@@ -1,0 +1,8 @@
+export interface HandlerOptions {
+  lang?: "ts" | "js";
+}
+
+export interface DirOptions {
+  events?: string;
+  commands?: string;
+}
